backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
httpx
langchain
langchain-groq
langchain-community
langchain-openai
chromadb
sqlalchemy
psycopg2-binary
supabase
pymupdf
python-docx
python-multipart
